hoge

+++ template.cpp
@@ -8,13 +8,14 @@
 #endif
 using namespace std;
 using i64 = long long;
-using pii = pair<int, int>;
+using pii = pair<i64, i64>;
 template<class A, class B>inline bool chmax(A &a, const B &b){return b>a ? a=b,1 : 0;}
 template<class A, class B>inline bool chmin(A &a, const B &b){return b<a ? a=b,1 : 0;}
 constexpr int INF  = 0x3f3f3f3f;
 constexpr i64 LINF = 0x3f3f3f3f3f3f3f3fLL;
 constexpr int MOD  = int(1e9) + 7;
 // >>>
+
 
 signed main()
 {
